Social Media Planner/backend/requirements.txt
flask==2.2.0
flask-sqlalchemy
flask-bcrypt
python-dotenv
flask-session>=0.4.0
redis
flask-cors
mysqlclient
Flask-Mail
requests
